backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
pymysql==1.1.1
pydantic[email]==2.9.2
python-multipart==0.0.9
PyJWT==2.9.0
cryptography
